+++ R/convert.R
@@ -140,32 +140,40 @@ to_imeca <- function(contaminant, value) {
 #'
 #' @param pollutant type of pollutant
 #' @param value value to convert to IMECAS
-#' @param showWarnings show a warning about issues with performing the conversion
+#' @param showWarnings Deprecated; use show_warnings instead.
+#' @param show_warnings show a warning about issues with performing the
+#' conversion
 #'
 #' @return value in IMECAS
 #' @export
 #' @examples
 #' ## IMECAs are a dimensionless scale that allows for the comparison of
 #' ## different pollutants
-#' convert_to_imeca(157, "O3", showWarnings = FALSE)
-#' convert_to_imeca(c(450, 350, 250), "NO2", showWarnings = FALSE)
-#' convert_to_imeca(80, "PM10", showWarnings = FALSE)
-#' convert_to_imeca(c(157, 200), "O3", showWarnings = FALSE)
-#' convert_to_imeca(48, "O3", showWarnings = FALSE)
-#' convert_to_imeca(67, "O3", showWarnings = FALSE)
-#' convert_to_imeca(77, "O3", showWarnings = FALSE)
-#' convert_to_imeca(205, "O3", showWarnings = FALSE)
-#' convert_to_imeca(72, "O3", showWarnings = FALSE)
-#' convert_to_imeca(98, "O3", showWarnings = FALSE)
+#' convert_to_imeca(157, "O3", show_warnings = FALSE)
+#' convert_to_imeca(c(450, 350, 250), "NO2", show_warnings = FALSE)
+#' convert_to_imeca(80, "PM10", show_warnings = FALSE)
+#' convert_to_imeca(c(157, 200), "O3", show_warnings = FALSE)
+#' convert_to_imeca(48, "O3", show_warnings = FALSE)
+#' convert_to_imeca(67, "O3", show_warnings = FALSE)
+#' convert_to_imeca(77, "O3", show_warnings = FALSE)
+#' convert_to_imeca(205, "O3", show_warnings = FALSE)
+#' convert_to_imeca(72, "O3", show_warnings = FALSE)
+#' convert_to_imeca(98, "O3", show_warnings = FALSE)
 #'
 #' ## Should show a warning because the conversion formula is not well
 #' ## defined
 #' convert_to_imeca(1.5, "CO")
 #'
-convert_to_imeca <- function(value, pollutant, showWarnings = TRUE) {
+convert_to_imeca <- function(value, pollutant, showWarnings = TRUE,
+                             show_warnings = TRUE) {
+  if (!missing("showWarnings")) {
+    warning("`showWarnings` argument deprecated. Use `show_warnings` instead.",
+            call. = FALSE)
+    show_warnings <- showWarnings
+  }
   pollutant <- toupper(pollutant)
   stopifnot(pollutant %in% c("O3", "NO2", "PM10", "SO2", "CO"))
-  if (showWarnings)
+  if (show_warnings)
     warning("This function is beta. Converted values don't always match official ones and care should be taken to validate them.")
   as.vector(unname(mapply(to_imeca, contaminant = pollutant, value = value)))
 }