+++ R/get_zone_imeca.R
@@ -1,8 +1,3 @@
-is.Date <- function(date, date.format = "%Y-%m-%d") {
-  tryCatch(!is.na(as.Date(date, date.format)),
-           error = function(e) {FALSE})
-}
-
 #' Title
 #'
 #' @param pollutant bla
@@ -17,7 +12,8 @@ is.Date <- function(date, date.format = "%Y-%m-%d") {
 #' @importFrom rvest html_nodes html_table
 #' @importFrom lubridate day month year
 #' @importFrom httr content
-.download_data_zone <- function(criterion, pollutant, zone, start_date, end_date) {
+.download_data_zone <- function(criterion, pollutant, zone, start_date,
+                                end_date) {
   url <- "http://www.aire.cdmx.gob.mx/estadisticas-consultas/consultas/resultado_consulta.php"
   fd <- list(
     diai	= day(start_date),
@@ -85,7 +81,8 @@ is.Date <- function(date, date.format = "%Y-%m-%d") {
 #' \strong{Zona Suroeste}: Álvaro Obregón,
 #' Coyoacán, Cuajimalpa, Magdalena Contreras, Tlalpan and Huixquilucan.
 #'
-#' @param pollutant The type of pollutant to download
+#' @param pollutant The type of pollutant to download. One or more of the
+#' following options:
 #' \itemize{
 #'  \item{"SO2"}{ - Dioxido de azufre}
 #'  \item{"CO"}{ - Monoxido de carbono}
@@ -94,7 +91,8 @@ is.Date <- function(date, date.format = "%Y-%m-%d") {
 #'  \item{"PM10"}{ - Particulas menores a 10 micrometros}
 #'  \item{"TC"}{- All the pollutants}
 #' }
-#' @param zone The geographic zone for which to download data
+#' @param zone The geographic zone for which to download data. One or more of
+#' the following:
 #' \itemize{
 #'  \item{"NO"}{ - Noroeste}
 #'  \item{"NE"}{ - Noreste}
@@ -103,16 +101,20 @@ is.Date <- function(date, date.format = "%Y-%m-%d") {
 #'  \item{"SE"}{ - Sureste}
 #'  \item{"TZ"}{ - All zones}
 #' }
-#' @param criterion The type of data to download
+#' @param criterion The type of data to download. One of the following options:
 #' \itemize{
 #'  \item{"HORARIOS"}{ - Hourly data}
 #'  \item{"MAXIMOS""}{ - Daily maximums}
 #' }
-#' @param start_date The start date in YYYY-MM-DD format (earliest possible value is 2008-01-01).
+#' @param start_date The start date in YYYY-MM-DD format (earliest possible
+#' value is 2008-01-01).
 #' @param end_date The end date in YYYY-MM-DD format.
-#' @param showWarnings Show warnings about problems with the data
+#' @param showWarnings Deprecated; use show_messages instead.
+#' @param show_messages show a message about issues with performing the
+#' conversion
 #'
-#' @return A data.frame with pollution data measured in IMECAS, by geographic zone. The hours
+#' @return A data.frame with pollution data measured in IMECAS, by geographic
+#' zone. The hours
 #' correspond to the \emph{Etc/GMT+6} timezone, with no daylight saving time
 #' @export
 #' @importFrom stringr str_c  str_replace_all
@@ -123,11 +125,11 @@ is.Date <- function(date, date.format = "%Y-%m-%d") {
 #' @examples
 #' ## There was a regional (NE) PM10 pollution emergency on Jan 6, 2017
 #' get_zone_imeca("MAXIMOS", "PM10", "NE", "2017-01-05", "2017-01-08",
-#'                showWarnings = FALSE)
+#'                show_messages = FALSE)
 #'
 #' ## There was an ozone pollution emergency on May 15, 2017
 #' get_zone_imeca("MAXIMOS", "O3", "TZ", "2017-05-15", "2017-05-15",
-#'                showWarnings = FALSE)
+#'                show_messages = FALSE)
 #'
 #' \dontrun{
 #' ## Download daily maximum PM10 data (particulate matter 10 micrometers or
@@ -147,41 +149,58 @@ is.Date <- function(date, date.format = "%Y-%m-%d") {
 #' }
 #'
 get_zone_imeca <- function(criterion, pollutant, zone, start_date, end_date,
-                          showWarnings = TRUE) {
+                          showWarnings = TRUE, show_messages = TRUE) {
+  if (!missing("showWarnings"))
+    warning("`showWarnings` argument deprecated.",
+            call. = FALSE)
   if (missing(pollutant))
-    stop("You need to specify a contaminante")
+    stop("You need to specify a pollutant")
   if (missing(zone))
     stop("You need to specify a zona")
   if (missing(criterion))
     stop("You need to specify a start date")
   if (missing(end_date))
-    stop("You need to specify an end date (YYYY-MM-DD)")
+    stop("You need to specify an end_date (YYYY-MM-DD)")
   if(!is.Date(end_date))
     stop("end_ate should be a date in YYYY-MM-DD format")
   if (missing(start_date))
-    stop("You need to specify a start date (YYYY-MM-DD)")
+    stop("You need to specify a start_date (YYYY-MM-DD)")
   if(!is.Date(start_date))
-    stop("start_date should be a date in YYYY-MM-DD format")
+    stop("start_date should be a string in YYYY-MM-DD format")
   if (start_date < "2008-01-01")
     stop("start_date should be after 2008-01-01")
 
   # standarize on uppercase since the station api expects upper, but
   # zone api expects lower
   criterion <- toupper(criterion)
-  stopifnot(length(base::setdiff(pollutant,
-                                 c("SO2", "CO", "NO2",
-                                   "O3", "PM10", "TC"))) == 0)
-  stopifnot(length(base::setdiff(zone,
-                                 c("NO", "NE", "CE", "SO", "SE", "TZ"))) == 0)
-  stopifnot(criterion %in% c("HORARIOS", "MAXIMOS"))
+  for (i in seq_len(length(pollutant)))
+    if (!(identical("O3", pollutant[i]) | identical("NO2", pollutant[i]) |
+          identical("SO2", pollutant[i]) | identical("CO", pollutant[i]) |
+          identical("PM10", pollutant[i]) | identical("TC", pollutant[i])))
+      stop("Invalid pollutant value")
+  pollutant <- unique(pollutant)
+  for (i in seq_len(length(zone)))
+    if (!(identical("NO", zone[i]) | identical("NE", zone[i]) |
+          identical("CE", zone[i]) | identical("SO", zone[i]) |
+          identical("SE", zone[i]) | identical("TZ", zone[i]) ))
+      stop("zone should be one of 'NO', 'NE', 'CE', 'SO', 'SE', or 'TZ'")
+  zone <- unique(zone)
+  if (!(identical("HORARIOS", criterion) | identical("MAXIMOS", criterion)))
+      stop("criterion should be 'HORARIOS' or 'MAXIMOS'")
   # the API expects lowercase letters
   criterion <- tolower(criterion)
 
-  # If pollutants are O3 or PM10 issue a warning that the way of calculating the index changed
-  if (length(base::intersect(pollutant, c("O3", "PM10"))) > 0 && showWarnings)
-    warning("\n*******************\nStarting October 28, 2014 the IMECA values for O3 and PM10 are computed using NOM-020-SSA1-2014 and NOM-025-SSA1-2014\n*******************")
-  if (start_date >= "2017-01-01" && showWarnings)
-    warning("\n*******************\nSometime in 2015-2017 the stations ACO, AJU, INN, MON, and MPA were excluded from the index\n*******************")
+  # If pollutants are O3 or PM10 display a message that the way of calculating
+  # the index changed
+  if (length(base::intersect(pollutant, c("O3", "PM10"))) > 0 && show_messages)
+    message(paste0("Starting October 28, 2014 the IMECA",
+                   " values for O3 and PM10 are computed using",
+                   " NOM-020-SSA1-2014 and",
+                   " NOM-025-SSA1-2014"))
+  if (start_date >= "2017-01-01" && show_messages)
+    message(paste0("Sometime in 2015-2017 the stations",
+                   " ACO, AJU, INN, MON, and MPA were excluded from the",
+                   " index"))
   df <- .download_data_zone(criterion, pollutant, zone, start_date, end_date)
 
   names(df) <- df[1, ]